Disable git status check on mounted volumes
When working with mounted volumes git runs really slow, and having a
status check in the prompt make the terminal insufferable to work with.

On OS X, the mounted volumes end up by default on `/Volumes`, so don't
show the status on the prompt when inside `/Volumes` altogether.

For now this is an OS X fix only--and will probably stay this way.